+++ cards/__init__.py
@@ -2,7 +2,7 @@
 
 """Top-level package for cards."""
 
-__version__ = '0.1.4'
+__version__ = '0.1.4.dev'  # developing 0.1.4
 
 from .cardsdb import (  # noqa: F401
     Card,