+++ docs/history.md
@@ -1,6 +1,13 @@
 History
 =======
 
+0.1.5 (2018-03-09)
+------------------
+- (internal) change version version handling in `setup.py` to read it from `cards/__init__.py`
+- (internal) take out versions from requirements_dev.txt
+- (internal) refacter tests directory into cli and cardsdb directories
+- (internal) finish tracer bullet tests for both API and CLI
+
 0.1.4 (2018-03-09)
 ------------------